do not catch and squelch async exception
This caused the following bug: a child thread, spawned by withAsync, and
linked to the main thread, threw an exception. The main thread was
presently trying to log something, so this simple logger grabbed the
async exception and tossed it out.

Solution: catch only IOExceptions.

This should also raise the question: if there's an exception in logging,
should we log it? i.e. should we have this exception handler at all? TBD